feat:  ajuste en egreso para que  permitir agregar decimales en el campo de pago

+++ src/app/modules/tesoreria/componentes/egreso/egreso-formulario/egreso-formulario.component.ts
@@ -220,9 +220,9 @@ export default class EgresoFormularioComponent
       const pago = detalleControl.get('pago')?.value || 0;
       const naturaleza = detalleControl.get('naturaleza')?.value;
       if (naturaleza === 'C') {
-        this.totalCredito += parseInt(pago);
+        this.totalCredito += parseFloat(pago);
       } else {
-        this.totalDebito += parseInt(pago);
+        this.totalDebito += parseFloat(pago);
       }
       this.changeDetectorRef.detectChanges();
     });
